Moved webxr.jsx to viewer/context
